@@ -40,7 +40,8 @@
     "clean": "rimraf lib dist es coverage",
     "api-types": "api-extractor run --local",
     "format": "prettier --write \"{src,test}/**/*.{js,ts,tsx}\" \"docs/**/*.md\"",
-    "lint": "eslint src --ext ts,tsx,js test/utils test/components test/hooks",
+    "lint": "eslint src test",
+    "lint:fix": "eslint src test --fix",
     "prepare": "yarn clean && yarn build",
     "pretest": "yarn lint",
     "test": "jest",
@@ -89,22 +90,22 @@
     "@testing-library/react-native": "^7.1.0",
     "@types/react": "18.2.25",
     "@types/react-native": "^0.67.4",
-    "@typescript-eslint/eslint-plugin": "^4.28.0",
-    "@typescript-eslint/parser": "^4.28.0",
+    "@typescript-eslint/eslint-plugin": "^6.17.0",
+    "@typescript-eslint/parser": "^6.17.0",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^29",
     "codecov": "^3.8.0",
     "cross-env": "^7.0.2",
-    "eslint": "^7.12.0",
-    "eslint-config-prettier": "^6.14.0",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.21.5",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-prettier": "^5.1.2",
+    "eslint-plugin-react": "^7.33.2",
     "glob": "^7.1.6",
     "jest": "^29",
     "jest-environment-jsdom": "^29.5.0",
     "metro-react-native-babel-preset": "^0.76.6",
-    "prettier": "^2.1.2",
+    "prettier": "^3.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "^0.71.11",
