@@ -7,12 +7,12 @@ import type { MapDispatchToPropsParam } from './selectorFactory'
 export function mapDispatchToPropsFactory<TDispatchProps, TOwnProps>(
   mapDispatchToProps:
     | MapDispatchToPropsParam<TDispatchProps, TOwnProps>
-    | undefined
+    | undefined,
 ) {
   return mapDispatchToProps && typeof mapDispatchToProps === 'object'
     ? wrapMapToPropsConstant((dispatch: Dispatch<Action<string>>) =>
         // @ts-ignore
-        bindActionCreators(mapDispatchToProps, dispatch)
+        bindActionCreators(mapDispatchToProps, dispatch),
       )
     : !mapDispatchToProps
     ? wrapMapToPropsConstant((dispatch: Dispatch<Action<string>>) => ({
