@@ -3,7 +3,7 @@ import { createInvalidArgFactory } from './invalidArgFactory'
 import type { MapStateToPropsParam } from './selectorFactory'
 
 export function mapStateToPropsFactory<TStateProps, TOwnProps, State>(
-  mapStateToProps: MapStateToPropsParam<TStateProps, TOwnProps, State>
+  mapStateToProps: MapStateToPropsParam<TStateProps, TOwnProps, State>,
 ) {
   return !mapStateToProps
     ? wrapMapToPropsConstant(() => ({}))
