@@ -15,7 +15,7 @@ export function createReduxContextHook(context = ReactReduxContext) {
 
     if (process.env.NODE_ENV !== 'production' && !contextValue) {
       throw new Error(
-        'could not find react-redux context value; please ensure the component is wrapped in a <Provider>'
+        'could not find react-redux context value; please ensure the component is wrapped in a <Provider>',
       )
     }
 
