@@ -2,7 +2,7 @@ import type { ActionCreatorsMapObject, Dispatch } from 'redux'
 
 export default function bindActionCreators(
   actionCreators: ActionCreatorsMapObject,
-  dispatch: Dispatch
+  dispatch: Dispatch,
 ): ActionCreatorsMapObject {
   const boundActionCreators: ActionCreatorsMapObject = {}
 
