@@ -4,23 +4,23 @@ describe('Utils', () => {
   describe('shallowEqual', () => {
     it('should return true if arguments fields are equal', () => {
       expect(
-        shallowEqual({ a: 1, b: 2, c: undefined }, { a: 1, b: 2, c: undefined })
+        shallowEqual({ a: 1, b: 2, c: undefined }, { a: 1, b: 2, c: undefined }),
       ).toBe(true)
 
       expect(shallowEqual({ a: 1, b: 2, c: 3 }, { a: 1, b: 2, c: 3 })).toBe(
-        true
+        true,
       )
 
       const o = {}
       expect(shallowEqual({ a: 1, b: 2, c: o }, { a: 1, b: 2, c: o })).toBe(
-        true
+        true,
       )
 
       const d = function () {
         return 1
       }
       expect(
-        shallowEqual({ a: 1, b: 2, c: o, d }, { a: 1, b: 2, c: o, d })
+        shallowEqual({ a: 1, b: 2, c: o, d }, { a: 1, b: 2, c: o, d }),
       ).toBe(true)
     })
 
@@ -40,8 +40,8 @@ describe('Utils', () => {
             d: function () {
               return 1
             },
-          }
-        )
+          },
+        ),
       ).toBe(false)
     })
 
@@ -57,8 +57,8 @@ describe('Utils', () => {
       expect(
         shallowEqual(
           { a: 1, b: 2, c: undefined },
-          { a: 1, bb: 2, c: undefined }
-        )
+          { a: 1, bb: 2, c: undefined },
+        ),
       ).toBe(false)
     })
 
